Two major updates: 1. Neighborhood Generation now takes the best k cells, not the first k cells. 2. Instead of having a fixed ratio of cells kept for all SimulatedAnnealing, each simulatedAnnealing has a variable percentage of kept cells.

+++ Solution.hpp
@@ -31,7 +31,8 @@ public:
 
     void addSolutionCell(SolutionCell *toAdd);
 
-    Solution *generateNeighbor(int kept, int *tasks, int sizeOfTasks, PeopleMatrix *people, CostMatrix *costs, int N);
+    Solution *
+    generateNeighbor(double kept, int *tasks, int sizeOfTasks, PeopleMatrix *people, CostMatrix *costs, int N);
 
     int getSize();
 
@@ -52,7 +53,7 @@ private:
     vector<SolutionCell> cells;
     PeopleMatrix *solutionPeople;
 
-    void populateSolution(Solution *initialSolution, int n, int *task, int sizeOfTasks, CostMatrix *costs, int N);
+    void populateSolution(Solution *initialSolution, double n, int *task, int sizeOfTasks, CostMatrix *costs, int N);
 
     int totalCost;
